refactor: msw - daily & streak 분리
