+++ src/mocks/data/streak/streakData.ts
@@ -1,0 +1,9 @@
+// 스트릭 캘린더 조회
+export const streakCalendarData = {
+    code: 'SUCCESS',
+    message: '요청에 성공했습니다.',
+    result: {
+      streakDateList: ['2025-06-10','2025-06-06', '2025-06-05', '2025-06-04', '2025-06-03', '2025-05-31', '2025-05-26'],
+    },
+  };
+
