+++ src/mocks/handlers/handlers.ts
@@ -2,5 +2,6 @@ import { authHandlers } from './auth';
 import { dailyHandlers } from './daily';
 import { fileHandlers } from './file';
 import { memberHandlers } from './member';
+import { streakHandlers } from './streak';
 
-export const handlers = [...authHandlers, ...dailyHandlers, ...memberHandlers, ...fileHandlers];
+export const handlers = [...authHandlers, ...dailyHandlers, ...memberHandlers, ...fileHandlers, ...streakHandlers];
